Switching to Commonjs module, how to include in README
