+++ index.html
@@ -12,7 +12,7 @@
 
   <link rel="stylesheet" type="text/css" href="../inorganik.github.io/assets/css/style.css?v=20171206">
 
-  <script data-main="./dist/demo.js" type="text/javascript" src="http://cdnjs.cloudflare.com/ajax/libs/require.js/2.1.1/require.min.js"></script>
+  <script src="./dist/demo-bundle.js"></script>
 	<script>
 
 	(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
